fix(ota): widen every mDNS A-record octet before shifting it
quint8(...) << 24 promotes the octet to int, and shifting a value >= 128
that far overflows a signed int. That is every address in 128.0.0.0/1 --
including the 192.168.x.x range a robot on a home network actually gets,
so the common case was the undefined one. It happened to produce the
right bits under the current compiler, which is why nothing caught it.

Widened to quint32 first, the idiom the rest of this codebase already
uses for exactly this (btphandshake.cpp:175, clocksync.cpp:52). Pinned
in test_mdnsresolver with 255.255.255.255 and the 128.0.0.1 boundary so
it cannot regress back into relying on the overflow.

+++ lib/ota/mdnsresolver.cpp
@@ -105,8 +105,15 @@ QHash<QString, QHostAddress> parseAAnswers(const QByteArray& message) {
         if (offset + rdlength > message.size()) break;
 
         if (type == 1 && rdlength == 4) {  // A record
-            const quint32 ip = (quint8(message.at(offset)) << 24) | (quint8(message.at(offset + 1)) << 16) |
-                                (quint8(message.at(offset + 2)) << 8) | quint8(message.at(offset + 3));
+            // quint32(quint8(...)), not a bare quint8: an octet promotes to
+            // int, and shifting a value >= 128 left by 24 overflows a signed
+            // int -- which is every address in 128.0.0.0/1, so every ordinary
+            // 192.168.x.x lease. Same idiom the rest of this codebase already
+            // uses to widen before shifting (btphandshake.cpp, clocksync.cpp).
+            const quint32 ip = (quint32(quint8(message.at(offset))) << 24) |
+                                (quint32(quint8(message.at(offset + 1))) << 16) |
+                                (quint32(quint8(message.at(offset + 2))) << 8) |
+                                quint32(quint8(message.at(offset + 3)));
             answers.insert(name, QHostAddress(ip));
         }
         offset += rdlength;
